+++ tests/system/test-api-search-post.js
@@ -150,6 +150,20 @@ test('/search sort', async (t) => {
   t.is(response.features[0].id, 'LC80100102015082LGN00')
 })
 
+test('/search sort unqualified field names fails', async (t) => {
+  const error = await t.throwsAsync(async () => t.context.api.client.post('search', {
+    json: {
+      sortby: [{
+        field: 'datetime',
+        direction: 'desc'
+      }]
+    }
+  }))
+
+  t.is(error.response.statusCode, 400)
+  t.truthy(error.response.body.description.includes('Hint: `sortby` requires fully qualified identifiers'))
+})
+
 test('/search flattened collection properties', async (t) => {
   let response = await t.context.api.client.post('search', {
     json: {
@@ -1614,4 +1628,111 @@ test('/search - context extension - context added when enabled', async (t) => {
   t.is(response.context.matched, 3)
   t.is(response.context.returned, 3)
   t.is(response.context.limit, 10)
+})
+
+test('/search invalid bbox throws error', async (t) => {
+  // test invalid longitude
+  {
+    const error = await t.throwsAsync(async () => t.context.api.client.post('search', {
+      json: {
+        bbox: [-190, -90, 180, 90]
+      }
+    }))
+    t.is(error.response.statusCode, 400)
+    t.is(error.response.body.code, 'BadRequest')
+    t.regex(
+      error.response.body.description,
+      // eslint-disable-next-line max-len
+      /Invalid \[lon, lat, lon, lat, z, z\] bbox\. {2}Longitudes must be between -180\/180, latitudes must be between {1}-90\/90, extent should not exceed \[-180, -90, 180, 90\]/
+    )
+  }
+
+  // test invalid latitude
+  {
+    const error = await t.throwsAsync(async () => t.context.api.client.post('search', {
+      json: {
+        bbox: [-110, -100, 180, 90]
+      }
+    }))
+    t.is(error.response.statusCode, 400)
+    t.is(error.response.body.code, 'BadRequest')
+    t.regex(
+      error.response.body.description,
+      // eslint-disable-next-line max-len
+      /Invalid \[lon, lat, lon, lat, z, z\] bbox\. {2}Longitudes must be between -180\/180, latitudes must be between {1}-90\/90, extent should not exceed \[-180, -90, 180, 90\]/
+    )
+  }
+
+  // test 6 coords with invalid values
+  {
+    const error = await t.throwsAsync(async () => t.context.api.client.post('search', {
+      json: {
+        bbox: [-190, -90, 180, 100, 10, 10]
+      }
+    }))
+    t.is(error.response.statusCode, 400)
+    t.is(error.response.body.code, 'BadRequest')
+    t.regex(
+      error.response.body.description,
+      // eslint-disable-next-line max-len
+      /Invalid \[lon, lat, lon, lat, z, z\] bbox\. {2}Longitudes must be between -180\/180, latitudes must be between {1}-90\/90, extent should not exceed \[-180, -90, 180, 90\]/
+    )
+  }
+})
+
+test('POST /search using "exclude" returns properly formatted links', async (t) => {
+  {
+    // test by excluding fields required to generate links ('id' and 'collection')
+    const response = await t.context.api.client.post(
+      'search',
+      {
+        resolveBodyOnly: false,
+        json: {
+          collections: ['landsat-8-l1'],
+          limit: 1,
+          fields: { exclude: ['assets', 'bbox', 'stac_version', 'id', 'collection'] }
+        } }
+    )
+    t.is(response.statusCode, 200)
+    t.is(response.body.features.length, 1)
+
+    const selfLink = response.body.features[0].links.find((l) => l.rel === 'self')
+    const selfPath = new URL(selfLink.href).pathname
+    t.is(selfPath, '/collections/landsat-8-l1/items/LC80100102015082LGN00')
+
+    const parentLink = response.body.features[0].links.find((l) => l.rel === 'parent')
+    const parentPath = new URL(parentLink.href).pathname
+    t.is(parentPath, '/collections/landsat-8-l1')
+
+    const collectionLink = response.body.features[0].links.find((l) => l.rel === 'collection')
+    const collectionPath = new URL(collectionLink.href).pathname
+    t.is(collectionPath, '/collections/landsat-8-l1')
+  }
+  {
+    // test without excluding fields required to generate links
+    const response = await t.context.api.client.post(
+      'search',
+      {
+        resolveBodyOnly: false,
+        json: {
+          collections: ['landsat-8-l1'],
+          limit: 1,
+          fields: { exclude: ['assets', 'bbox', 'stac_version'] }
+        } }
+    )
+    t.is(response.statusCode, 200)
+    t.is(response.body.features.length, 1)
+
+    const selfLink = response.body.features[0].links.find((l) => l.rel === 'self')
+    const selfPath = new URL(selfLink.href).pathname
+    t.is(selfPath, '/collections/landsat-8-l1/items/LC80100102015082LGN00')
+
+    const parentLink = response.body.features[0].links.find((l) => l.rel === 'parent')
+    const parentPath = new URL(parentLink.href).pathname
+    t.is(parentPath, '/collections/landsat-8-l1')
+
+    const collectionLink = response.body.features[0].links.find((l) => l.rel === 'collection')
+    const collectionPath = new URL(collectionLink.href).pathname
+    t.is(collectionPath, '/collections/landsat-8-l1')
+  }
 })